# openapi.yaml
openapi: 3.0.3
info:
  title: JobMajunga (Efiploi) API
  version: 1.0.0
  description: API REST centralisée (Node.js/Express/MySQL) pour WPF + Flutter.
servers:
  - url: https://api.jobmajunga.local/v1
    description: Local/Dev (exemple)
...
security:
  - bearerAuth: []
paths:
...
  /applications:
    get:
      tags: [Applications]
      summary: Liste candidatures (candidat = les siennes, recruteur = sur ses offres)
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [sent, viewed, reviewing, interview, accepted, rejected]
        - in: query
          name: jobOfferId
          schema: { type: integer, minimum: 1 }
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: pageSize
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        "200":
          description: Liste paginée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedApplications"
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [Applications]
      summary: Postuler à une offre (candidat)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ApplicationCreate" }
      responses:
        "201":
          description: Candidature créée
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Application" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
  /applications/{id}:
    get:
      tags: [Applications]
      summary: Détail candidature (candidat propriétaire ou recruteur propriétaire de l’offre)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        "200":
          description: Candidature
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Application" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /applications/{id}/status:
    patch:
      tags: [Applications]
      summary: Changer statut (recruteur ou admin, progression seulement)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, minimum: 1 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [viewed, reviewing, interview, accepted, rejected]
      responses:
        "200":
          description: Candidature mise à jour
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Application" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /applications/{id}/notes:
    patch:
      tags: [Applications]
      summary: Notes internes recruteur (recruteur propriétaire)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, minimum: 1 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [recruiterNotes]
              properties:
                recruiterNotes: { type: string, nullable: true }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Application" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /applications/{id}/interview:
    patch:
      tags: [Applications]
      summary: Planifier entretien (recruteur propriétaire)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, minimum: 1 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [interviewDate]
              properties:
                interviewDate:
                  type: string
                  format: date-time
                location:
                  type: string
                  nullable: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Application" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /applications/{id}/interview-response:
    patch:
      tags: [Applications]
      summary: Accepter ou refuser un entretien (candidat propriétaire)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, minimum: 1 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [response]
              properties:
                response:
                  type: string
                  enum: [accepted, refused]
      responses:
        "200":
          description: Réponse enregistrée
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Application" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Requête invalide
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    Forbidden:
      description: Accès interdit (RBAC)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Ressource introuvable
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: Conflit (ex. déjà existant / règle métier)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      required: [message]
      properties:
        message: { type: string }
        code: { type: string, nullable: true }
...
    Application:
      type: object
      required: [id, candidateId, jobOfferId, cvId, status, appliedAt, updatedAt]
      properties:
        id: { type: integer }
        candidateId: { type: integer }
        jobOfferId: { type: integer }
        cvId: { type: integer }
        coverLetter: { type: string, nullable: true }
        status:
          type: string
          enum: [sent, viewed, reviewing, interview, accepted, rejected]
        recruiterNotes: { type: string, nullable: true }
        interviewDate: { type: string, format: date-time, nullable: true }
        candidateInterviewStatus:
          type: string
          enum: [pending, accepted, refused]
        appliedAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    ApplicationCreate:
      type: object
      required: [jobOfferId, cvId]
      properties:
        jobOfferId: { type: integer, minimum: 1 }
        cvId: { type: integer, minimum: 1 }
        coverLetter: { type: string, nullable: true }
    PaginatedApplications:
      type: object
      required: [items, page, pageSize, total]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Application" }
        page: { type: integer }
        pageSize: { type: integer }
        total: { type: integer }
...